backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.26.0
python-dotenv==1.0.1
groq==0.4.2
google-api-python-client==2.111.0
